Update install.sql
Edited install.sql to be compatible with mysql server versions on linux the IF NOT EXISTS clause is not compatible inside of CREATE TABLE or ALTER TABLE aswell as an issue with the use of curtime() as a default value for the createdAt column. MySQL does not support using a function as a default value for a DATETIME column.